fix: update register schema zod validations
